webservice/ncei_wind/requirements.txt
Flask==1.0.2
google-cloud-pubsub
google-cloud-bigquery
requests
